+++ .gitignore
@@ -1,0 +1,6 @@
+__pycache__/
+*.pyc
+*.pyo
+*.egg-info/
+*.3mf
+*.obj
