+++ INSTRUCTIONS_RU.md
@@ -1,0 +1,24 @@
+# Инструкция по запуску
+
+1. Установите Python 3.10 или новее.
+2. Установите зависимости и сам пакет командой:
+
+   ```bash
+   pip install -e .
+   ```
+
+3. Для запуска GUI выполните:
+
+   ```bash
+   python -m grid_split --gui
+   ```
+
+4. Для запуска из командной строки используйте:
+
+   ```bash
+   python -m grid_split --input MODEL.STL --size 250 250 250 --merged out.3mf
+   ```
+
+Флаг `--repair` включает попытку починки модели, `--fallback` выбирает режим
+резервного пересечения (`planar` по умолчанию).
+
