+++ requirements.txt
@@ -1,0 +1,3 @@
+trimesh
+numpy
+pyglet<2
